app/requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.14.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
db-dtypes>=1.2.0
openai>=1.0.0
